refactor: split toolsProvider into focused modules + fix 3 bugs
## What changed

### Bug fixes
- `search_directory`: fix broken regex escape (`'\\$&'` was corrupted to
  a string containing literal source code, so literal-string searches
  never matched anything)
- `change_directory`: was using `validatePath()` which blocked navigation
  to parent dirs or absolute paths; restored original `resolve()` behaviour
- `getPersistedState`: `defaultWorkspacePath` config was silently ignored
  whenever a persisted state file existed; now the explicit config always
  wins so the plugin starts in the folder the user actually configured

### Refactor — src/tools/ (new)
Split the 3 787-line `toolsProvider.ts` into focused, single-responsibility
modules. `toolsProvider.ts` is now a 103-line orchestrator.

| File | Responsibility |
|---|---|
| `context.ts` | Shared mutable `ToolContext` type |
| `helpers.ts` | `validatePath`, `createSafeToolImplementation`, RAG utils |
| `fileTools.ts` | All file I/O, search, memory |
| `codeTools.ts` | JS/Python execution, shell, terminal, background commands |
| `webTools.ts` | Web search (multi-provider), fetch, RAG, Wikipedia |
| `browserTools.ts` | Puppeteer session open/control/close, one-shot open |
| `gitTools.ts` | git status/diff/commit/log/add/checkout |
| `githubTools.ts` | gh auth/issues/PRs/push |
| `miscTools.ts` | System info, clipboard, open/preview, documents, DB, local RAG |
| `subAgentTools.ts` | Full secondary-agent loop with auto-save and debug pass |

All 93 existing tests pass. Smoke-tested in a live LM Studio instance.
